// generated/assembly/dsp-op.ts
// TypeScript + AssemblyScript Ops Enum
// auto-generated from scripts
export enum Op {
  End,
  Begin,
  CreateGen,
  CreateAudios,
  CreateValues,
  AudioToScalar,
  LiteralToAudio,
  Pick,
  Pan,
  SetValue,
  SetValueDynamic,
  SetProperty,
  UpdateGen,
  ProcessAudio,
  ProcessAudioStereo,
  BinaryOp
}